ctr_api/pyproject.toml
[tool.poetry]
name = "ctr-api"
version = "0.1.0"
description = ""
authors = ["dedoxd2 <dedo.xd33@yahoo.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.2"
djangorestframework = "^3.14.0"
python-decouple = "^3.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
